+++ Detectors/EMCAL/reconstruction/CMakeLists.txt
@@ -16,6 +16,9 @@ o2_add_library(EMCALReconstruction
                        src/RawHeaderStream.cxx
                        src/RAWDataHeader.cxx
                        src/AltroDecoder.cxx
+                       src/RCUTrailer.cxx
+                       src/Bunch.cxx
+                       src/Channel.cxx
                        src/Mapper.cxx
                PUBLIC_LINK_LIBRARIES FairRoot::Base O2::Headers
                                      AliceO2::InfoLogger
@@ -27,6 +30,9 @@ o2_target_root_dictionary(
                           HEADERS include/EMCALReconstruction/RawReaderFile.h
                                   include/EMCALReconstruction/RawReaderMemory.h
                                   include/EMCALReconstruction/AltroDecoder.h
+                                  include/EMCALReconstruction/RCUTrailer.h
+                                  include/EMCALReconstruction/Bunch.h
+                                  include/EMCALReconstruction/Channel.h
                                   include/EMCALReconstruction/RAWDataHeader.h
                                   include/EMCALReconstruction/Mapper.h)
