+++ Detectors/EMCAL/reconstruction/src/Bunch.cxx
@@ -1,0 +1,18 @@
+// Copyright CERN and copyright holders of ALICE O2. This software is
+// distributed under the terms of the GNU General Public License v3 (GPL
+// Version 3), copied verbatim in the file "COPYING".
+//
+// See http://alice-o2.web.cern.ch/license for full licensing information.
+//
+// In applying this license CERN does not waive the privileges and immunities
+// granted to it by virtue of its status as an Intergovernmental Organization
+// or submit itself to any jurisdiction.
+#include "EMCALReconstruction/Bunch.h"
+
+using namespace o2::emcal;
+
+void Bunch::initFromRange(gsl::span<uint16_t> adcs)
+{
+  for (auto adcval : adcs)
+    mADC.emplace_back(adcval);
+}
