+++ Detectors/EMCAL/reconstruction/src/EMCALReconstructionLinkDef.h
@@ -14,11 +14,15 @@
 #pragma link off all classes;
 #pragma link off all functions;
 
-#pragma link C++ class o2::emcal::RawReaderFile < o2::emcal::RAWDataHeader> + ;
-#pragma link C++ class o2::emcal::RawReaderFile < o2::header::RAWDataHeaderV4> + ;
-#pragma link C++ class o2::emcal::RawReaderMemory < o2::emcal::RAWDataHeader> + ;
+#pragma link C++ class o2::emcal::RawReaderFile < o2::emcal::RAWDataHeader > +;
+//#pragma link C++ class o2::emcal::RawReaderFile < o2::header::RAWDataHeaderV4> + ;
+#pragma link C++ class o2::emcal::RawReaderMemory < o2::emcal::RAWDataHeader > +;
 //#pragma link C++ class o2::emcal::RawReaderMemory < o2::header::RAWDataHeaderV4> + ;
-#pragma link C++ class o2::emcal::AltroDecoder + ;
+#pragma link C++ class o2::emcal::AltroDecoder < o2::emcal::RawReaderFile < o2::emcal::RAWDataHeader >> +;
+#pragma link C++ class o2::emcal::AltroDecoder < o2::emcal::RawReaderMemory < o2::emcal::RAWDataHeader >> +;
+#pragma link C++ class o2::emcal::Bunch + ;
+#pragma link C++ class o2::emcal::Channel + ;
+#pragma link C++ class o2::emcal::RCUTrailer + ;
 #pragma link C++ class o2::emcal::Mapper + ;
 
 #endif
