+++ Detectors/EMCAL/reconstruction/src/RawReaderFile.cxx
@@ -129,4 +129,4 @@ void RawReaderFile<RawHeader>::readFile(const std::string_view filename)
 }
 
 template class o2::emcal::RawReaderFile<o2::emcal::RAWDataHeader>;
-template class o2::emcal::RawReaderFile<o2::header::RAWDataHeaderV4>;
+//template class o2::emcal::RawReaderFile<o2::header::RAWDataHeaderV4>;
